requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pytube>=12.1.0
moviepy>=1.0.3
python-multipart
python-dotenv

# pydantic==1.10.13
pdfplumber==0.11.0
python-docx==0.8.11
python-pptx==0.6.23
pytesseract==0.3.10

requests==2.31.0

beautifulsoup4>=4.12.0

html2text>=2020.1.16

youtube-transcript-api==0.6.1